refactor: re-organize files
