+++ packages/account-tree-controller/src/group.ts
@@ -1,11 +1,11 @@
 import { type AccountGroup, type AccountGroupId } from '@metamask/account-api';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
 
-import type { AccountTreeWallet } from './AccountTreeWallet';
 import type {
   AccountGroupObject,
   AccountTreeControllerMessenger,
 } from './types';
+import type { AccountTreeWallet } from './wallet';
 
 export const DEFAULT_ACCOUNT_GROUP_NAME: string = 'Default';
